requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Settings
pydantic-settings==2.9.1
python-dotenv==1.0.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
aiosqlite==0.20.0
alembic==1.14.1

# OpenAI
openai==1.82.0

# Audio processing
pydub==0.25.1
noisereduce==3.0.3
numpy==1.26.4
scipy==1.14.1

# Transcription & Diarization
whisperx==3.3.1
faster-whisper==1.1.0
torch>=2.0.0
torchaudio>=2.0.0

# Job queue
redis==5.2.1
arq==0.26.1

# Utilities
aiofiles==24.1.0
python-jose==3.3.0
httpx==0.28.1

# PDF export
reportlab==4.2.5
